Realizados cambios en el CRUD del proveedor

+++ src/main/java/com/storecode/controllers/ProviderController.java
@@ -49,13 +49,17 @@ public class ProviderController {
 	}
 	
 	@PostMapping("/updateProvider/{idProvider}")
-	public String updateProvider(@PathVariable("idProvider") long idProvider, Provider provider,
-			BindingResult fields, Model model){
+	public String updateProvider(@PathVariable("idProvider") long idProvider,
+			@ModelAttribute("provider") @Valid Provider provider, BindingResult fields, Model model){
 		if (fields.hasErrors()) {
 			provider.setId(idProvider);
 			return "provider/updateProvider";
 		}else {
-			providerService.save(provider);
+			Provider providerAux = providerService.getById(idProvider);
+			providerAux.setName(provider.getName());
+			providerAux.setUbication(provider.getUbication());
+			providerAux.setPhone(provider.getPhone());
+			providerService.save(providerAux);
 			model.addAttribute("listProviders", providerService.getAll());
 			return "redirect:/providers/listProviders";
 		}
@@ -63,9 +67,15 @@ public class ProviderController {
 	
 	@GetMapping("/deleteProvider/{idProvider}")
     public String deleteProvider(@PathVariable("idProvider") long idProvider, Model model) {
-        providerService.deleteById(idProvider);
-        model.addAttribute("providers", providerService.getAll());
-        return "redirect:/providers/listProviders";
+		Provider provider = providerService.getById(idProvider);
+		if(providerService.existsProductByProvider(provider)){
+			model.addAttribute("error", "El proveedor no puede ser eliminado porque se encuentra con productos a cargo");
+			return "redirect:/providers/listProviders";
+		}else {
+			providerService.deleteById(idProvider);
+	        model.addAttribute("providers", providerService.getAll());
+	        return "redirect:/providers/listProviders";
+		}
     }
 	
 	@GetMapping("/listProviders")
